Extract skill path details
